Add server entities, repo, and endpoints

+++ backend/src/main/java/com/yapper/repository/ServerRepository.java
@@ -3,5 +3,7 @@ package com.yapper.repository;
 import com.yapper.entity.Server;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServerRepository extends JpaRepository<Server, Long> {
+public interface ServerRepository extends JpaRepository<Server, Integer> {
+
+
 }
